requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
redis==3.5.3
requests==2.25.0
simplejson==3.17.2
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
